Move ReactDebugCurrentFiber to 'shared' directory
**what is the change?:**
See title

**why make this change?:**
We need this in many of our packages, and some were already requiring it
using the full path. Let's just move it to 'shared'.

**test plan:**
Ran build and tests.

**issue:**
Internal task T30481686
