+++ packages/react-dom/src/client/ReactDOMFiberInput.js
@@ -7,8 +7,7 @@
  * @flow
  */
 
-// TODO: direct imports like some-package/src/* are bad. Fix me.
-import ReactDebugCurrentFiber from 'react-reconciler/src/ReactDebugCurrentFiber';
+import ReactDebugCurrentFiber from 'shared/ReactDebugCurrentFiber';
 import invariant from 'shared/invariant';
 import warning from 'shared/warning';
