+++ packages/react-dom/src/client/ReactDOMFiberTextarea.js
@@ -9,8 +9,7 @@
 
 import invariant from 'shared/invariant';
 import warning from 'shared/warning';
-// TODO: direct imports like some-package/src/* are bad. Fix me.
-import ReactDebugCurrentFiber from 'react-reconciler/src/ReactDebugCurrentFiber';
+import ReactDebugCurrentFiber from 'shared/ReactDebugCurrentFiber';
 
 import ReactControlledValuePropTypes from '../shared/ReactControlledValuePropTypes';
