+++ packages/react-reconciler/src/ReactFiberContext.js
@@ -17,7 +17,7 @@ import invariant from 'shared/invariant';
 import warning from 'shared/warning';
 import checkPropTypes from 'prop-types/checkPropTypes';
 
-import ReactDebugCurrentFiber from './ReactDebugCurrentFiber';
+import ReactDebugCurrentFiber from 'shared/ReactDebugCurrentFiber';
 import {startPhaseTimer, stopPhaseTimer} from './ReactDebugFiberPerf';
 import {createCursor, push, pop} from './ReactFiberStack';
